[be] refactor : 테스트용 파일 생성 삭제 하는 부분 중복 코드 제거
테스트용 파일 생성과 삭제를 맡는 유틸 클래스를 생성하여 중복코드 제거
