+++ src/test/java/me/hoyoung/youtube/domain/video/VideoDriveTest.java
@@ -1,24 +1,17 @@
 package me.hoyoung.youtube.domain.video;
 
+import me.hoyoung.youtube.util.MockFile;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
-import java.io.FileInputStream;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Arrays;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -30,36 +23,25 @@ public class VideoDriveTest {
     @Autowired
     private VideoDrive videoDrive;
 
-    @Value("${video.storage.dir}")
-    private String directory;
-
-    private String testExtension = ".tests";
-
-    @BeforeEach
-    public void createDir() {
-        File dir = new File(directory);
-        if (!dir.exists()) {
-            dir.mkdirs();
-        }
-    }
+    @Autowired
+    private MockFile mockFile;
 
     @AfterEach
     public void cleanUp() {
-        final File[] files = new File(directory).listFiles((dir, name) -> name.matches( ".+("+ testExtension +")$" ));
-        Arrays.stream(files).forEach((file) -> file.delete());
+        mockFile.cleanUp();
     }
 
     @Test
     @DisplayName("비디오 createFile 테스트")
     public void createFileTest() throws Exception {
         //given
-        MultipartFile multipartFile = createTestRandomFile("Test contents");
+        MultipartFile multipartFile = mockFile.createRandomFile("Test contents");
 
         //when
         String createdFileName = videoDrive.createFile(multipartFile);
 
         //then
-        File createFile = new File(directory + createdFileName);
+        File createFile = mockFile.findFile(createdFileName);
         assertThat(createFile.exists()).isTrue();
     }
 
@@ -68,7 +50,7 @@ public class VideoDriveTest {
     public void getFileTest() throws Exception {
         //given
         String contents = UUID.randomUUID().toString() + "TestTestTestTestTestTest";
-        MultipartFile multipartFile =  createTestRandomFile(contents);
+        MultipartFile multipartFile =  mockFile.createRandomFile(contents);
         String createdFileName = videoDrive.createFile(multipartFile);
 
         //when
@@ -77,14 +59,6 @@ public class VideoDriveTest {
         //then
         String readContents = com.google.common.io.Files.asCharSource(getFile, StandardCharsets.UTF_8).read();
         assertThat(readContents).isEqualTo(contents);
-    }
-
-    public MultipartFile createTestRandomFile(String contents) throws Exception {
-        String filename = UUID.randomUUID().toString() + testExtension;
-        byte[] contentBytes = contents.getBytes();
-        Path path = Paths.get(directory + filename);
-        Files.write(path, contentBytes);
-        return new MockMultipartFile(filename, filename, "text/plain",  new FileInputStream(new File(directory + filename)));
     }
 
 }
