+++ src/test/java/me/hoyoung/youtube/web/VideoControllerTest.java
@@ -3,14 +3,13 @@ package me.hoyoung.youtube.web;
 import me.hoyoung.youtube.domain.user.User;
 import me.hoyoung.youtube.domain.user.UserRepository;
 import me.hoyoung.youtube.domain.video.VideoRepository;
+import me.hoyoung.youtube.util.MockFile;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
@@ -19,14 +18,6 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.io.File;
-import java.io.FileInputStream;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Arrays;
-import java.util.UUID;
 
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
@@ -47,10 +38,9 @@ public class VideoControllerTest {
     @Autowired
     private WebApplicationContext context;
 
-    @Value("${video.storage.dir}")
-    private String directory;
+    @Autowired
+    private MockFile mockFile;
 
-    private String testExtension = ".tests";
     private MockMvc mvc;
     private UserDetails userDetails;
 
@@ -69,30 +59,17 @@ public class VideoControllerTest {
     public void tearDown() {
         videoRepository.deleteAll();
         userRepository.deleteAll();
-        cleanUp();
-    }
-
-    public void cleanUp() {
-        final File[] files = new File(directory).listFiles((dir, name) -> name.matches( ".+("+ testExtension +")$" ));
-        Arrays.stream(files).forEach((file) -> file.delete());
+        mockFile.cleanUp();
     }
 
     @Test
     @WithMockUser(roles = "USER")
     public void videoSaveApiTest() throws Exception {
-        MultipartFile mockVideo = createTestRandomFile("videoSaveAPI TEST");
+        MultipartFile mockVideo = mockFile.createRandomFile("videoSaveAPI TEST");
 
         mvc.perform(MockMvcRequestBuilders.multipart("/api/v1/video/upload")
                 .file("content", mockVideo.getBytes())
                 .with(user(userDetails)))
                 .andExpect(status().isOk());
-    }
-
-    public MultipartFile createTestRandomFile(String contents) throws Exception {
-        String filename = UUID.randomUUID().toString() + testExtension;
-        byte[] contentBytes = contents.getBytes();
-        Path path = Paths.get(directory + filename);
-        Files.write(path, contentBytes);
-        return new MockMultipartFile(filename, filename, "text/plain",  new FileInputStream(new File(directory + filename)));
     }
 }
